[C#] move all files under Solution  - Add DiffMatchPatch project  - Add DiffMatchPatchTest project for xUnit tests
