+++ csharp/DiffMatchPatch.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiffMatchPatch", "DiffMatchPatch\DiffMatchPatch.csproj", "{13DE99BE-E8C4-496D-8489-E2C10C0AA10B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiffMatchPatch.Tests", "DiffMatchPatch.Tests\DiffMatchPatch.Tests.csproj", "{CB1CA713-68C4-49DB-A62E-987B74E891DB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{13DE99BE-E8C4-496D-8489-E2C10C0AA10B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{13DE99BE-E8C4-496D-8489-E2C10C0AA10B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{13DE99BE-E8C4-496D-8489-E2C10C0AA10B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{13DE99BE-E8C4-496D-8489-E2C10C0AA10B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB1CA713-68C4-49DB-A62E-987B74E891DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB1CA713-68C4-49DB-A62E-987B74E891DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB1CA713-68C4-49DB-A62E-987B74E891DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB1CA713-68C4-49DB-A62E-987B74E891DB}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
