+++ csharp/DiffMatchPatch/DiffMatchPatch.csproj
@@ -1,0 +1,6 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <RootNamespace>Google.DiffMatchPatch</RootNamespace>
+  </PropertyGroup>
+</Project>
